+++ CHANGELOG.rst
@@ -11,8 +11,9 @@ New Features
 Breaking changes
 ~~~~~~~~~~~~~~~~
 - Simulation campaigns, circuits, and reports using BlueConfig format aren't supported anymore.
-- The analysis configuration accepts population and node_set, instead of target.
-- Rename the function ``blueetl.core.utils.safe_concat`` to ``smart_concat``.
+- The analysis configuration accepts ``population`` and ``node_set``, instead of ``target``.
+- In the ``neuron_classes`` definition, the query parameters must be moved to ``query``, ``$limit`` must be renamed to ``limit``, ``$gids`` to ``node_id``.
+- The function ``blueetl.core.utils.safe_concat`` has been renamed to ``smart_concat``.
 
 Improvements
 ~~~~~~~~~~~~